fix: hide text in listen template on AnkiDroid
AnkiDroid shows the text that it is going to speak per default. This is okay for the back side of cards, but it defeats the purpose for the listen cards. Do some mild hacking to ensure it is hidden on AnkiDroid, but also that the desktop version does not show any unwanted text.