fix: forbid code splitting
Stop a deduplication feature of parcel kicking in. It loads an external script to avoid duplicated code, which is cool for most situations, but we really really need a single file without external scripts, so we set the threshold of necessary duplications for the feature to kick in to some impossible high value. With this, we again have no external script files.