feat: generate apkg with build
Generate an apkg with vocabulary that uses the template on each build. It is based on a skeleton file with some placeholder templates for the actual templates to be patched in during build. To customize the vocabulary later, users can modify their anki collection and then use that as the new skeleton. For this to work, the HTML files have been modified to contain unique identifiers that are also present in the placeholder templates. In order to be able to check in a minimal version again, the build also generates a new skeleton file to be checked in. Actual releases will probably not include the skeleton prominently, because it should not be used directly. The problem is that it would even work for learning, but it is ugly - users should only use the rich apkg with the templates.