fix: no speech output for hanzi templates
When TTS is enabled in AnkiDroid, it reads all the text on the card out loud unless there is a `tts` tag. Add such a tag with no content, so that no speech output happens, which is not the point of these particular two templeates (only learning Hanzi and their meanings).

Issue: #8