feat: add new Q/A type for non-language facts
Add the new card type, integrate it into the build and index.html. Also change the layout for all cards a bit so that the card type on the top floats, and the main content clears. For this purpose a new class .prompt has been added to the translations, as well as the new card type. The new facts type additionally shows the subdeck on the right for context.